+++ apollo-composition/src/lib.rs
@@ -82,7 +82,7 @@ pub trait HybridComposition {
                         .locations
                         .into_iter()
                         .map(|range| SubgraphLocation {
-                            subgraph: subgraph.name.clone(),
+                            subgraph: Some(subgraph.name.clone()),
                             range: Some(range),
                         })
                         .collect(),
@@ -133,23 +133,16 @@ pub trait HybridComposition {
                 self.update_supergraph_sdl(raw_sdl);
                 let satisfiability_result = self.validate_satisfiability().await;
                 self.add_issues(
-                    satisfiability_result_into_issues(satisfiability_result)
-                        .map(|mut issue| {
-                            for (service_name, connector) in by_service_name.iter() {
-                                issue.message = issue.message.replace(
-                                    &**service_name,
-                                    connector.id.subgraph_name.as_str(),
-                                );
-                            }
-                            issue
-                        })
-                        .chain(once(Issue {
-                            code: "EXPERIMENTAL_FEATURE".to_string(),
-                            message: "Connectors are an experimental feature. Breaking changes are likely to occur in future versions.".to_string(),
-                            locations: vec![],
-                            severity: Severity::Warning,
-                        })),
+                    satisfiability_result_into_issues(satisfiability_result).map(|mut issue| {
+                        for (service_name, connector) in by_service_name.iter() {
+                            issue.message = issue
+                                .message
+                                .replace(&**service_name, connector.id.subgraph_name.as_str());
+                        }
+                        issue
+                    }),
                 );
+
                 self.update_supergraph_sdl(original_supergraph_sdl);
             }
             ExpansionResult::Unchanged => {
@@ -181,7 +174,10 @@ pub struct Issue {
 /// A location in a subgraph's SDL
 #[derive(Clone, Debug)]
 pub struct SubgraphLocation {
-    pub subgraph: String,
+    /// This field is an Option to support the lack of subgraph names in
+    /// existing composition errors. New composition errors should always
+    /// include a subgraph name.
+    pub subgraph: Option<String>,
     pub range: Option<Range<LineColumn>>,
 }
 
@@ -270,7 +266,7 @@ impl From<Issue> for BuildMessage {
 impl From<SubgraphLocation> for BuildMessageLocation {
     fn from(location: SubgraphLocation) -> Self {
         BuildMessageLocation {
-            subgraph: Some(location.subgraph),
+            subgraph: location.subgraph,
             start: location.range.as_ref().map(|range| BuildMessagePoint {
                 line: Some(range.start.line),
                 column: Some(range.start.column),
@@ -292,7 +288,7 @@ impl From<SubgraphLocation> for BuildMessageLocation {
 impl SubgraphLocation {
     fn from_ast(node: SubgraphASTNode) -> Option<Self> {
         Some(Self {
-            subgraph: node.subgraph?,
+            subgraph: node.subgraph,
             range: node.loc.and_then(|node_loc| {
                 Some(Range {
                     start: LineColumn {
@@ -402,7 +398,7 @@ impl From<BuildHint> for Issue {
 impl From<BuildMessageLocation> for SubgraphLocation {
     fn from(location: BuildMessageLocation) -> Self {
         Self {
-            subgraph: location.subgraph.unwrap_or_default(),
+            subgraph: location.subgraph,
             range: location.start.and_then(|start| {
                 let end = location.end?;
                 Some(Range {