+++ apollo-composition/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## vNEXT
+
+- [#586](https://github.com/apollographql/federation-rs/pull/586) Make
+  `SubgraphLocation.subgraph` an `Option`. For now, composition errors can have
+  no attributed subgraph.
+
 ## 0.1.2
 
-* Updated dependencies
+- Updated dependencies